==> Notebooks/04-XGBoost_Course_Prepare_Iris_Dataset.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn import preprocessing

<h2>Iris Classification Dataset</h2>

Input Features:<br>
sepal_length,sepal_width,petal_length,petal_width<br>

Target:<br>
Iris plant class<br>

Objective: Predict iris plant class for a given sepal_length,sepal_width,petal_length,petal_width<br>
<h4>Data source: https://archive.ics.uci.edu/ml/datasets/iris</h4>

columns = ['encoded_class','sepal_length','sepal_width','petal_length','petal_width']

# Encode Class Labels to integers
le = preprocessing.LabelEncoder()
le.fit(['Iris-setosa', 'Iris-versicolor', 'Iris-virginica'])

le.classes_

df = pd.read_csv('../Data/iris_all.csv')

df['class'].value_counts()

df.head()

df.tail()

le.transform(df['class'])[-5:]

# Convert Classes to numeric value
df['encoded_class'] = le.transform(df['class'])

df.head()

df.tail()

# Visualize
setosa = df['class'] == 'Iris-setosa'
versicolor = df['class'] == 'Iris-versicolor'
virginica = df['class'] == 'Iris-virginica'

plt.scatter(df[setosa].sepal_length,y=df[setosa].sepal_width, label='setosa',color='g')
plt.scatter(df[versicolor].sepal_length,y=df[versicolor].sepal_width, label='versicolor',color='r')
plt.scatter(df[virginica].sepal_length,y=df[virginica].sepal_width, label='virginica',color='b')
plt.xlabel('length')
plt.ylabel('width')
plt.title('Sepal')
plt.grid(True)
plt.legend()
plt.show()

plt.scatter(df[setosa].petal_length,y=df[setosa].petal_width, label='setosa',color='g')
plt.scatter(df[versicolor].petal_length,y=df[versicolor].petal_width, label='versicolor',color='r')
plt.scatter(df[virginica].petal_length,y=df[virginica].petal_width, label='virginica',color='b')
plt.xlabel('length')
plt.ylabel('width')
plt.title('Petal')
plt.grid(True)
plt.legend()
plt.show()

plt.scatter(df[setosa].petal_length,y=df[setosa].sepal_length, label='setosa',color='g')
plt.scatter(df[versicolor].petal_length,y=df[versicolor].sepal_length, label='versicolor',color='r')
plt.scatter(df[virginica].petal_length,y=df[virginica].sepal_length, label='virginica',color='b')
plt.xlabel('petal length')
plt.ylabel('sepal length')
plt.title('Petal-Sepal')
plt.grid(True)
plt.legend()
plt.show()

## Training and Validation Set
### Target Variable as first column followed by input features:
class,sepal_length,sepal_width,petal_length,petal_width
### Training, Validation files do not have a column header

# Training = 70% of the data
# Validation = 30% of the data
# Randomize the datset
np.random.seed(5)
l = list(df.index)
np.random.shuffle(l)
df = df.iloc[l]

rows = df.shape[0]
train = int(.7 * rows)
test = rows-train

rows, train, test

# Write Training Set
df[:train].to_csv('../Data/iris_train.csv'
                          ,index=False,header=False
                          ,columns=columns)

# Write Validation Set
df[train:].to_csv('../Data/iris_validation.csv'
                          ,index=False,header=False
                          ,columns=columns)

# Write Column List
with open('../Data/iris_train_column_list.txt','w') as f:
    f.write(','.join(columns))
